on same file upload error fixed

+++ src/utils/cloudinary.js
@@ -8,22 +8,24 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-const uploadOnCloudinary = async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath, removeFile = true) => {
   try {
     if (!localFilePath) return null
     //upload the file on cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto"
     })
-    // file has been uploaded successfull
-    
-    //both files have to be different or this will give error that " Error: ENOENT: no such file or directory, unlink  "
-    
-    fs.unlinkSync(localFilePath)
+    // file has been uploaded successfully
+
+    if (removeFile) {
+      fs.unlinkSync(localFilePath)
+    }
     return response;
 
   } catch (error) {
-    fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+    if (removeFile) {
+      fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+    }
     return null;
   }
 }
